requirements.txt
aiohttp==3.6.2
alabaster==0.7.12
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.14.62
botocore==1.17.62
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cssselect2==0.3.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.10
django-appconf==1.0.4
django-compressor==2.4
django-countable-field==1.3
django-countries==6.1.3
django-crispy-forms==1.9.2
django-crontab==0.7.1
django-db-mailer==2.4.0
django-discover-runner==1.0
django-easy-audit==1.3.0
django-environ==0.4.5
django-excel==0.0.10
django-ipware==3.0.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-postgres-copy==2.4.2
django-redis==4.12.1
django-reversion==3.0.8
django-reversion-compare==0.12.2
django-role-permissions==3.1.0
django-ses==1.0.3
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
Faker==4.1.3
future==0.18.2
geoip2==4.0.2
gunicorn==20.0.4
html5lib==1.1
httpagentparser==1.9.0
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
iniconfig==1.0.1
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.2
isoweek==1.3.3
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
lml==0.0.9
lxml==4.5.2
MarkupSafe==1.1.1
maxminddb==2.0.2
mccabe==0.6.1
more-itertools==8.5.0
multidict==4.7.6
mws==0.8.10
mypy-extensions==0.4.3
openpyxl==3.0.5
packaging==20.4
parso==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipdeptree==1.0.0
pluggy==0.13.1
prompt-toolkit==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyexcel==0.6.4
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.9
pyexcel-xlsx==0.5.8
pyflakes==2.2.0
Pygments==2.7.0
pylint==2.6.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
Pyphen==0.9.5
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
rcssmin==1.0.6
redis==3.5.3
requests==2.24.0
rjsmin==1.1.0
s3transfer==0.3.3
six==1.15.0
snowballstemmer==2.0.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
Sphinx==3.2.1

sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.3
tinycss2==1.0.2
toml==0.10.1
traitlets==5.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
WeasyPrint==51
webencodings==0.5.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
yarl==1.5.1
zipp==3.1.0
